Add files via upload
These are Arduino applications and a java source code that allow for many types of sensors such as RGB, ultrasonic, and more to read in recorded values from the sensors, output the data on the screen and perform functions based on the inputted data. 

The LCD clock Java source code and Arduino application is a project that uses the Serial Port library to communicate data inputted from the arduino serial ports from the Uno Rev 3 board and outputs data into an Array that is used in a Java Gui to show the connection and outputs. 
The ardunio circuit is connected to an LCD screen that will display the output of the java gui and specifically the current date and time. IT will also allow for input to be inserted from the user through the keyboard and print out onto the LCD screen. 

The Light ultrasonic Arduino source codes, are for a circuit that uses the HC-SR04 Ultrasonic sensor connected to an Arduino Uno board using a Arduino Uno rev 3 motor shield on top. Two dc motors are connected to the motor shield, allowing for programmable control of the speed, direction and power of each individual motors. An ISL 29125 breakout RGB light sensor from sparkfun is connected to the motor shield to allow for Light variants to be inputted and displayed through the output terminal. This code then uses loops as well as the specific library functions to determine the distance, and values of light being inputted. I then use these values in selection statements to determine the action of the motors dependent upon the color RGB values as well as distance and object values determined.  If the RGB reads in a red light the motors will slow down and turn off simulating a stop. If green is read in and determined the motors will speed up and maintain that speed, blue(yellow) will simulate the motors slowing down.  

The same version with the avoidance attaches an avoidance sensor similar to a motion sensor that will allow the vehicle to stop once the sensor reads in an object. This sensor is much more simplistic then the ultrasonic as its goal is to only read in an object if its present or not this will allow the motors to come to a stop immediately if needed with this sensor.